fix: create common function for getting event id
