+++ packages/trpc/server/routers/viewer/eventTypes.tsx
@@ -7,14 +7,9 @@ import { z } from "zod";
 import getAppKeysFromSlug from "@calcom/app-store/_utils/getAppKeysFromSlug";
 import { DailyLocationType } from "@calcom/app-store/locations";
 import { stripeDataSchema } from "@calcom/app-store/stripepayment/lib/server";
-import { StripeData } from "@calcom/app-store/stripepayment/lib/server";
-import { getEventTypeAppData, getLocationOptions } from "@calcom/app-store/utils";
-import { LocationObject } from "@calcom/core/location";
-import { parseBookingLimit, parseRecurringEvent, validateBookingLimitOrder } from "@calcom/lib";
-import getEnabledApps from "@calcom/lib/apps/getEnabledApps";
+import { validateBookingLimitOrder } from "@calcom/lib";
 import { CAL_URL } from "@calcom/lib/constants";
-import getStripeAppData from "@calcom/lib/getStripeAppData";
-import { getTranslation } from "@calcom/lib/server/i18n";
+import getEventTypeById from "@calcom/lib/getEventTypeById";
 import { baseEventTypeSelect, baseUserSelect } from "@calcom/prisma";
 import { _DestinationCalendarModel, _EventTypeModel } from "@calcom/prisma/zod";
 import {
@@ -469,232 +464,14 @@ export const eventTypesRouter = router({
         throw new TRPCError({ code: "INTERNAL_SERVER_ERROR" });
       }
 
-      const userSelect = Prisma.validator<Prisma.UserSelect>()({
-        name: true,
-        username: true,
-        id: true,
-        avatar: true,
-        email: true,
-        locale: true,
-        defaultScheduleId: true,
+      const res = await getEventTypeById({
+        eventTypeId: input.id,
+        userId: ctx.user.id,
+        prisma: ctx.prisma,
+        isTrpcCall: true,
       });
 
-      const rawEventType = await ctx.prisma.eventType.findFirst({
-        where: {
-          AND: [
-            {
-              OR: [
-                {
-                  users: {
-                    some: {
-                      id: ctx.session.user.id,
-                    },
-                  },
-                },
-                {
-                  team: {
-                    members: {
-                      some: {
-                        userId: ctx.session.user.id,
-                      },
-                    },
-                  },
-                },
-                {
-                  userId: ctx.session.user.id,
-                },
-              ],
-            },
-            {
-              id: input.id,
-            },
-          ],
-        },
-        select: {
-          id: true,
-          title: true,
-          slug: true,
-          description: true,
-          length: true,
-          hidden: true,
-          locations: true,
-          eventName: true,
-          customInputs: true,
-          timeZone: true,
-          periodType: true,
-          metadata: true,
-          periodDays: true,
-          periodStartDate: true,
-          periodEndDate: true,
-          periodCountCalendarDays: true,
-          requiresConfirmation: true,
-          recurringEvent: true,
-          hideCalendarNotes: true,
-          disableGuests: true,
-          minimumBookingNotice: true,
-          beforeEventBuffer: true,
-          afterEventBuffer: true,
-          slotInterval: true,
-          hashedLink: true,
-          bookingLimits: true,
-          successRedirectUrl: true,
-          currency: true,
-          team: {
-            select: {
-              id: true,
-              slug: true,
-              members: {
-                where: {
-                  accepted: true,
-                },
-                select: {
-                  role: true,
-                  user: {
-                    select: userSelect,
-                  },
-                },
-              },
-            },
-          },
-          users: {
-            select: userSelect,
-          },
-          schedulingType: true,
-          schedule: {
-            select: {
-              id: true,
-            },
-          },
-          userId: true,
-          price: true,
-          destinationCalendar: true,
-          seatsPerTimeSlot: true,
-          seatsShowAttendees: true,
-          workflows: {
-            include: {
-              workflow: {
-                include: {
-                  activeOn: {
-                    select: {
-                      eventType: {
-                        select: {
-                          id: true,
-                          title: true,
-                        },
-                      },
-                    },
-                  },
-                  steps: true,
-                },
-              },
-            },
-          },
-        },
-      });
-
-      if (!rawEventType) {
-        throw new TRPCError({ code: "NOT_FOUND" });
-      }
-
-      const credentials = await ctx.prisma.credential.findMany({
-        where: {
-          userId: ctx.session.user.id,
-          app: {
-            enabled: true,
-          },
-        },
-        select: {
-          id: true,
-          type: true,
-          key: true,
-          userId: true,
-          appId: true,
-          invalid: true,
-        },
-      });
-
-      const { locations, metadata, ...restEventType } = rawEventType;
-      const newMetadata = EventTypeMetaDataSchema.parse(metadata || {})!;
-      const apps = newMetadata.apps || {};
-      const eventTypeWithParsedMetadata = { ...rawEventType, metadata: newMetadata };
-      newMetadata.apps = {
-        ...apps,
-        stripe: {
-          ...getStripeAppData(eventTypeWithParsedMetadata, true),
-          currency:
-            (
-              credentials.find((integration) => integration.type === "stripe_payment")
-                ?.key as unknown as StripeData
-            )?.default_currency || "usd",
-        },
-        giphy: getEventTypeAppData(eventTypeWithParsedMetadata, "giphy", true),
-        rainbow: getEventTypeAppData(eventTypeWithParsedMetadata, "rainbow", true),
-      };
-
-      // TODO: How to extract metadata schema from _EventTypeModel to be able to parse it?
-      // const parsedMetaData = _EventTypeModel.parse(newMetadata);
-      const parsedMetaData = newMetadata;
-
-      const parsedCustomInputs = (rawEventType.customInputs || []).map((input) =>
-        customInputSchema.parse(input)
-      );
-
-      const eventType = {
-        ...restEventType,
-        schedule: rawEventType.schedule?.id || rawEventType.users[0]?.defaultScheduleId || null,
-        recurringEvent: parseRecurringEvent(restEventType.recurringEvent),
-        bookingLimits: parseBookingLimit(restEventType.bookingLimits),
-        locations: locations as unknown as LocationObject[],
-        metadata: parsedMetaData,
-        customInputs: parsedCustomInputs,
-      };
-
-      // backwards compat
-      if (eventType.users.length === 0 && !eventType.team) {
-        const fallbackUser = await ctx.prisma.user.findUnique({
-          where: {
-            id: ctx.session.user.id,
-          },
-          select: userSelect,
-        });
-        if (!fallbackUser)
-          throw new TRPCError({
-            code: "NOT_FOUND",
-            message: "The event type doesn't have user and no fallback user was found",
-          });
-        eventType.users.push(fallbackUser);
-      }
-      const currentUser = eventType.users.find((u) => u.id === ctx.session.user.id);
-      const t = await getTranslation(currentUser?.locale ?? "en", "common");
-      const integrations = await getEnabledApps(credentials);
-      const locationOptions = getLocationOptions(integrations, t);
-
-      const eventTypeObject = Object.assign({}, eventType, {
-        periodStartDate: eventType.periodStartDate?.toString() ?? null,
-        periodEndDate: eventType.periodEndDate?.toString() ?? null,
-      });
-
-      const teamMembers = eventTypeObject.team
-        ? eventTypeObject.team.members.map((member) => {
-            const user = member.user;
-            user.avatar = `${CAL_URL}/${user.username}/avatar.png`;
-            return user;
-          })
-        : [];
-
-      // Find the current users memebership so we can check role to enable/disable deletion.
-      // Sets to null if no membership is found - this must mean we are in a none team event type
-      const currentUserMembership =
-        eventTypeObject.team?.members.find((el) => el.user.id === ctx.session.user.id) ?? null;
-
-      const finalObj = {
-        eventType: eventTypeObject,
-        locationOptions,
-        team: eventTypeObject.team || null,
-        teamMembers,
-        currentUserMembership,
-      };
-      return finalObj;
+      return res;
     }),
   update: eventOwnerProcedure.input(EventTypeUpdateInput.strict()).mutation(async ({ ctx, input }) => {
     const {
